Dance Website- Taking user Info through(Mongoose)
In this project we have used mongoose, In order to Get the contact data from the users. Similarly, we can also take the sign up and login details from any user visiting our website. 

We have Used the Pug Template Engine and The Extending the template in the other pug files using the 'extends' and 'block' keywords. We have used express as the backend to render the static files and using mongoose. We have used get request and post request [Exploring the express.js]